refactor: dto 파일 구조 변경 - controller 패키지 내부로 이동
